refactor: isolate chat pagination boundary helper
